requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
openai>=1.0.0
pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.1.0
python-dotenv>=1.0.0
faker>=20.0.0
reportlab>=4.0.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
googlemaps>=4.10.0
werkzeug>=3.0.0

